refactor oauth sign in actions
